pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "larry-recsys-data-foundation"
version = "0.1.0"
description = "Data foundation (Overture + OSM conflation) for RPG TDD POC"
requires-python = ">=3.11"
dependencies = [
  "pytest>=8.0.0",
  "pandas>=2.2.0",
  "duckdb>=1.0.0",
  "pyarrow>=15.0.0",
  "apache-airflow>=2.9.0",
  "google-cloud-storage>=2.17.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

